Fix bug in blocked svd and evd
The blocked implementations operate on matrices of size 2 * block size
by 2 * block size. If the matrix size is 2 * block size by 2 * block
size or smaller then there is no point in applying blocking. Previously
we considered only the case where the matrix is smaller than 2 * block
size by 2 * block size and not equal, which leads to unnecessary
computations that do not add anything to the final result.